add simple and trivial array class

+++ src/prg.hpp
@@ -274,7 +274,7 @@ namespace pr {
 		void visitPolyMesh(std::function<void(std::shared_ptr<const FPolyMeshEntity>)> visitor) const { visit(visitor); }
 		void visitPolyMesh(std::function<VisitorAction(std::shared_ptr<const FPolyMeshEntity>)> visitor) const { visit(visitor); }
 
-		int entityCount() const { return _entities.size(); }
+		int entityCount() const { return (int)_entities.size(); }
 		std::shared_ptr<FSceneEntity> entityAt(int i) { return _entities[i]; }
 	private:
 		std::vector<std::shared_ptr<FSceneEntity>> _entities;
